Refactor preferences
